Allow non-Unicode named groups
This patch implements Kevin Gibbons' disambiguation of syntax using
Waldemar Horwat's proposed strategy of reparsing at the end if a
named group is found. See #7 for more background.

Note that a practical implementation could switch [N] mode on on-line
and only reparse from the beginning in the unlikely case of a \k
backreference preceding a named group, based on remembering whether
the unlikely sequence \k has occurred.