Normative: Remove complicated SyntaxError from Replacement
Previously, if the letters enclosed in "$<...>" did not form an
IdentifierName, a SyntaxError was thrown. However, this is rather
complicated to implement, and the semantics otherwise would be
reasonable: to replace it with getting that value from groups.
If a bulit-in RegExp is used, this will always result in the
empty string.