Normative: Put the groups object as the third-to-last replace argument
Currently, the last two arguments of the replace function have
a certain value, and the previous arguments may be of variable length
based on how many captures are included in the RegExp. If someone is
using String.prototype.replace with a user-provided RegExp, rather than
a particular set one, then they might depend on the invariant that the
last two arguments have those values.

To maintain that invariant, this patch puts the groups object as the
third-to-last, rather than the last, argument for the functional
replacer. Users may need to change their replace function if they
change their RegExp to use named groups to account for this, but
they already need to change the function to accomodate more arguments
if they add more groups, so these semantics may be analogous.

This change is based on a suggestion from several committee members at
the January 2017 TC39 meeting.