Revert "Normative: Put the groups object as the third-to-last replace argument"
This reverts commit 95e91a9f3cd1a91edc84e4de3c64685fd5524c4b.

The revert here was based on further discussion at the March 2017 TC39 meeting.
It was judged that the compatibility impact is likely to be more minimal with
this argument order than the other, as RegExps passed to `replace` tend to
be literal, and the last two arguments may be used positionally.